Moved/fixed location of method
